move styles + remove padding
