+++ src/App.tsx
@@ -13,7 +13,6 @@ const App = () => {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    padding: 10,
   },
 });
