+++ src/components/screens/CommunityMembersScreen/CommunityMembersScreen.styles.tsx
@@ -1,0 +1,17 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  search: {
+    paddingBottom: 10,
+  },
+  noResultContainer: {
+    flex: 1,
+    alignItems: 'center',
+    padding: 50,
+  },
+  noResultText: {
+    fontSize: 20,
+  },
+});
+
+export default styles;
